refactor(subgraph): split into package with construction/validation/management and re-exports
Move subgraph implementation into `src/meridian/core/subgraph/` preserving API (expose_input/output, inputs_of, node_names, validate). Remove monolithic file. Tests pass.
