+++ src/meridian/core/subgraph/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from .construction import Subgraph
+
+__all__ = ["Subgraph"]
