+++ src/meridian/core/subgraph/execution.py
@@ -1,0 +1,3 @@
+from __future__ import annotations
+
+# Placeholder for future subgraph execution helpers if needed.
