+++ src/meridian/core/subgraph/management.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+from .construction import Subgraph, ValidationIssue
+
+
+def validate_subgraph_structure(g: Subgraph) -> list[ValidationIssue]:
+    return g.validate()
